apis/notebook-to-api/requirements.txt
fastapi
python-dotenv
langchain
qdrant
pydantic
uvicorn
